refactor to sentiment file
